requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
PyJWT==2.10.1
bcrypt==4.3.0
python-multipart==0.0.6
email-validator==2.1.0
